# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/isvalid": {
            "get": {
                "description": "是否登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "是否登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "userid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户邮箱",
                        "name": "email",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-model_TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    }
                }
            }
        },
        "/refresh": {
            "post": {
                "description": "刷新登录令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "刷新登录令牌",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户Refresh Token",
                        "name": "refreshToken",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-model_TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errhandle.ErrCode": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                6,
                7,
                8,
                9,
                10,
                11,
                12,
                13,
                14
            ],
            "x-enum-varnames": [
                "NoError",
                "InnerError",
                "TokenError",
                "ParamsError",
                "UserExists",
                "UserNonExists",
                "PasswordInvalid",
                "PermissionDenied",
                "ClassError",
                "SexError",
                "NameExists",
                "EmailFormatError",
                "RealNameFormatError",
                "SchoolError",
                "PasswordTooShort"
            ]
        },
        "model.CommonResponse-any": {
            "type": "object",
            "properties": {
                "data": {},
                "reason": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/errhandle.ErrCode"
                }
            }
        },
...
        "model.CommonResponse-model_TokenResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.TokenResponse"
                },
                "reason": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/errhandle.ErrCode"
                }
            }
        },
...
        "model.TokenResponse": {
            "type": "object",
            "properties": {
                "expiredAt": {
                    "type": "integer"
                },
                "refreshtoken": {
                    "type": "string"
                },
                "scope": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        }
    }
}